Split part of UnitTestWidget off in new TestRunner class
